pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isaqe"
version = "0.1.0"
description = "Intelligent Schema-Aware Query Executor"
authors = [{name = "Codex"}]
requires-python = ">=3.9"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "asyncpg",
  "redis",
  "httpx",
  "tenacity",
  "sqlglot",
  "pandas",
  "numpy",
  "jinja2",
  "python-dateutil",
  "pyyaml",
  "jsonschema",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-fastapi",
  "prometheus-client",
  "structlog",
  "scikit-learn",
  "orjson"
]

[tool.pyright]
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "respx",
  "locust"
]
